# docs/specs/api-openapi.yaml
...
openapi: 3.1.0
info:
  title: PromptPotter Control-remote API
  version: 0.0.0-profile-minus-1
  summary: Closed inbound command set for the M12 Control-remote I/O kind.
  description: |
    The HTTP command surface for PromptPotter's interactive control plane.
    Every operation is a Control-remote command: an HTTP-ingressed mutation
    authored by a signed-in operator or signed-in client, appended to its
    target ledger (the per-cycle `.runtime/ledger.jsonl` or the workspace
    `.workspace/events.jsonl`, per command scope) as a `CommandRecord` by the
    sole `CommandDispatcher`.

    See `docs/adr/0001-m12-control-plane.md` for the permanent contract, the
    Profile gradient (A-E), and the 20-item security checklist that every
    handler is measured against.

    Conformance: handlers MUST reject any request that omits the
    `Idempotency-Key` header. State-mutating handlers MUST additionally
    reject any request that omits the `Expected-Version` header. Mismatch
    on `Expected-Version` returns 409 `version_conflict`.

  contact:
    name: PromptPotter maintainers
  license:
    name: See repository LICENSE
servers:
  - url: http://localhost:8001
    description: Local dev (uvicorn + auth-off, Stage 0 / Profile B default).
  - url: https://{install}.promptpotter.com
    description: Hosted hub (Profile D — placeholder, OIDC required).
    variables:
      install:
        default: example
        description: Per-install hostname slug.
...
paths:
...
  /commands/fork-cycle:
    post:
      operationId: forkCycle
      summary: Fork from the parent's current ledger offset; retargets active pointer.
      description: |
        Rides `CycleEventLog.inherit_from(parent, offset)` via the unified
        `mint_operator_fork` path. `round` / `candidate_id` name the searchpoint;
        the typed `ForkSpec` lands on the parent's `FORK_CUT` ledger record and
        the fork's `index.json::fork`. Every operator fork is `operator_steered`:
        the required `seed` roots the fork at the chosen searchpoint plus the
        operator's prompt/config edits and reconciled run limits. After minting,
        the fork is **launched** (an implicit `start-run`/resume against the new
        cycle), so steer-and-fork continues optimizing from the edited searchpoint
        in one gesture — the reconciled limits in `seed` govern its run.
      tags: [cycle-lifecycle]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ExpectedVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CommandEnvelope'
                - type: object
                  required: [kind, payload]
                  properties:
                    kind: { const: fork-cycle }
                    payload: { $ref: '#/components/schemas/ForkCyclePayload' }
      responses:
        '202': { $ref: '#/components/responses/CommandAccepted' }
        '400': { $ref: '#/components/responses/IdempotencyKeyMissing' }
        '403': { $ref: '#/components/responses/CapabilityDenied' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/PayloadInvalid' }
...
  /campaigns:
    get:
      operationId: listCampaigns
      summary: Every campaign in one store owned by the caller, newest first.
      description: |
        The forest's other half — `GET /cycles` returns one store's cycles,
        this returns its campaigns, and the webapp groups the pair by
        `root_cycle_id` (`buildForest`). Takes the same `descend`: absent/empty
        is the caller's own tenant tree; a hop chain re-roots the list into an
        L4 `.inner/<key>` sandbox, so one route lists the top level and
        any depth. Response shape is owned by the server model
        (`routers/campaigns/manifests.py::CampaignListResponse`, generated into
        `webapp/lib/api/types.generated.ts::CampaignSummary`) — not duplicated
        here. Each summary carries the campaign's lifetime billed spend
        (`spend_used_usd`, a floor while `spend_unpriced_tokens > 0`), summed
        server-side over its cycle ledgers and banked tombstones — the
        per-campaign split of `GET /auth/quota-status`'s account total. Beside
        it, `spend_metered` is what the campaign's cap counts along its line —
        the bill, or a controlled arm's search incurred — by bucket: the
        number the sidebar prints, so it never picks or adds one up itself.
        Each carries `bench`, the headline read off the campaign's result (`campaigns/{id}/result.json`,
        `domain/campaign.py::CampaignResult`) under the formula its line runs —
        whichever cycle a rebase handed the line to — and never a per-cycle copy.
        Each also carries
        `runs_with` — what the ROOT course runs with (models and every scalar
        setting, each with the layer that won it, the optimizer manifest it
        selects, plus the DECLARED rounds cap)
        — a second transport of the `GET /campaigns/{id}/pipeline` answer at the
        root, never a second source; null when the root pipeline did not
        resolve, so one broken dataset drops its own row's answer only.
      tags: [cycles]
      parameters:
        - name: dataset
          in: query
          required: false
          description: Filter to one dataset's campaigns.
          schema: { type: string }
        - name: lifecycle
          in: query
          required: false
          description: |
            `active` (default), `archived`, `deleted` — the visibility intent —
            or `all`. `checkin` is not a status but the authoring PHASE, read off
            the root cycle's `.runtime/checkin.flag`; it narrows `active`, which
            already carries the origin-authoring set.
          schema:
            type: string
            enum: [active, archived, deleted, checkin, all]
            default: active
        - name: descend
          in: query
          required: false
          description: |
            `~`-joined `campaign::cycle` hops to descend into, as on
            `GET /cycles`. Empty/absent → the caller's own tree.
          schema: { type: string }
      responses:
        '200':
          description: The resolved store's campaigns, newest first (`CampaignListResponse`).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: A `descend` hop names an inner sandbox that does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /campaigns/{campaign_id}/checkin:
    get:
      operationId: getCampaignCheckin
      summary: Re-open a CHECKIN campaign — its draft wire + last resolver turn.
      description: |
        The sidebar opens a check-in campaign through here instead of
        the dashboard (no `dashboard.json` exists pre-loop): the webapp rebuilds
        the ingest "ready" panel from `draft` and shows the prior `resolution`
        recap. Tenant-scoped — a cross-tenant id 404s. 404 when the campaign has
        no check-in working state (already Started or never a check-in).
      tags: [campaigns]
      parameters:
        - name: campaign_id
          in: path
          required: true
          description: The check-in campaign id (the draft's `draft_id`).
          schema:
            type: string
            minLength: 8
            maxLength: 128
      responses:
        '200':
          description: |
            The rehydrated draft wire + the last resolver turn (or null) + any
            proposals the last turn left unclicked.
          content:
            application/json:
              schema:
                type: object
                required: [draft, resolution, raised]
                properties:
                  draft: { $ref: '#/components/schemas/DraftCampaign' }
                  resolution:
                    description: |
                      The last `resolve-origin` turn's output (assessment,
                      findings, next_action, recap), or null when none ran.
                    type: [object, 'null']
                  raised:
                    type: array
                    items: { $ref: '#/components/schemas/RaisedCommand' }
                    description: |
                      Proposals the last turn left unclicked. Always present
                      (`[]` when none) — without them a re-opened check-in would
                      drop the operator's outstanding actions on the floor.
                    additionalProperties: true
        '404':
          description: No check-in working state for this campaign (or cross-tenant).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /campaigns/{campaign_id}/pipeline:
    get:
      operationId: getCampaignPipeline
      summary: What this campaign's pipeline RESOLVES TO at one searchpoint.
      description: |
        The one server-owned answer to "what config does this node run" — the
        sole source for every config row, model chip and provenance badge
        (`frontend-surface-contract.md::I9`). Five layers, merged
        last-writer-wins by `application/pipeline_resolve.py`: dataset overlay <
        the campaign's FROZEN `pipeline_overlay` < cycle seed < the addressed
        candidate's sparse `pipeline_overlay` < identity contributions,
        which are unoverridable. Each param's `source` is the layer that won it,
        stamped BY the merge. A check-in campaign resolves through the same
        call. A one-shot read beside `dashboard.json` polling, which is its
        invalidation signal, not its transport (`roadmap.md` § Hard ordering).
      tags: [campaigns]
      parameters:
        - name: campaign_id
          in: path
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
        - name: at
          in: query
          required: false
          description: |
            Which searchpoint to resolve, in the ONE subject grammar
            (`application/evidence/subjects.py::parse_subject`): `campaign:<id>`,
            `course:<id>/<cycle>/<label>`, `candidate:<id>/<cycle>/<cand_id>`,
            each optionally suffixed `;in=<hop chain>` for an L4 inner point.
            Omitted defaults to `campaign:<campaign_id>`. `;lens=` and
            `;samples=` are REFUSED with 400: a scoring mask cannot change what
            config a point ran. An `at` naming another campaign is 400 too.
          schema:
            type: string
            maxLength: 512
      responses:
        '200':
          description: The resolved pipeline at the addressed searchpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignPipelineResponse'
        '400':
          description: |
            Malformed `at`, a scoring mask on a configuration read, or an `at`
            naming a campaign other than the path's.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: |
            No such campaign, or cross-tenant. Ownership-gated by `load_owned`
            and nothing else — 404 rather than 403, so a probe cannot enumerate
            other tenants' campaign ids.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID-or-equivalent identifying this command attempt.
        The dispatcher dedupes by hash lookup against the target ledger; a
        second POST with the same `Idempotency-Key` returns the original 202
        response, not a fresh command.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9_-]+$'
    ExpectedVersion:
      name: Expected-Version
      in: header
      required: false
      description: |
        The ledger offset (or round number, for round-grained commands) the
        client believes the target cycle is at. The runner rejects the command
        with 409 `version_conflict` if the cycle has advanced past this point.
        Optimistic-concurrency floor.

        **v0 relaxation (Phase 2b):** the header is declared optional in v0
        because the webapp's mutation surface does not yet thread ledger
        sequence numbers from the SSE stream through to mutations. The
        dispatcher validates the version when the header IS present and
        skips the check otherwise. Phase 2c flips this back to `required:
        true` once the webapp consumes sequence from `EventStreamView`.
      schema:
        type: integer
        minimum: 0
  schemas:
...
    CommandEnvelope:
      type: object
      required: [kind, payload]
      additionalProperties: false
      properties:
        kind:
          type: string
          description: |
            The command kind. MUST be one of the kinds declared in this
            file's `paths` keys (the closed set). Examples once Profile B
            lands: `pause-cycle`, `fork-cycle`,
            `change-run-limits`. See `docs/adr/0001-m12-control-plane.md`
            Profile B table.
          minLength: 1
          maxLength: 64
          pattern: '^[a-z][a-z0-9-]*$'
        payload:
          description: |
            Per-kind payload. Each declared operation in `paths` overrides
            this with a tighter schema (using `$ref` or inline). The base
            envelope only guarantees the field is present.
          type: object
          additionalProperties: true
...
  responses:
...
    IdempotencyKeyMissing:
      description: Request omitted the mandatory `Idempotency-Key` header.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: pp_session
      description: |
        Opaque server-side session id (NOT a JWT). Issued by the OIDC
        middleware at Profile C after ID Token verification; per
        identity-foundation no-drift gate #2 the cookie itself carries no
        claims. At Profile B (auth-off), this scheme is declared but the
        Stage-0 default identity is returned unconditionally.